edited getAll of voucher

+++ Application/LineItems/Queries/GetLineItem/LineItemDto.cs
@@ -7,11 +7,11 @@ namespace StockManagment.Application.LineItems.Queries.GetLineItem
     {
         public string Code { get; set; }
         public string Remark { get; set; }
-        public double UnitAmount { get; set; }
-        public double Quantity { get; set; }
-        public double TaxableAmount { get; set; }
+        public float UnitAmount { get; set; }
+        public float Quantity { get; set; }
+        public float TaxableAmount { get; set; }
         public string TaxType { get; set; } = "VAT";
-        public double TaxAmount { get; set; }
+        public float TaxAmount { get; set; }
         public string ElementCode { get; set; }
         public string VoucherCode { get; set; }
     }
