+++ Application/Vouchers/Commands/UpdateVoucher/UpdateVoucherCommand.cs
@@ -16,8 +16,8 @@ namespace StockManagment.Application.Vouchers.Commands.UpdateVoucher
         public string Remark { get; set; }
         public string Type { get; set; }
         public DateTime TimeStamp { get; set; }
-        public double SubTotal { get; set; }
-        public double GrandTotal { get; set; }
+        public float SubTotal { get; set; }
+        public float GrandTotal { get; set; }
         public bool Void { get; set; }
         public string ConsigneeCode { get; set; }
     }
