+++ Domain/Entities/LineItem.cs
@@ -4,13 +4,12 @@ using System.Collections.Generic;
 namespace StockManagment.Domain.Entities
 {
     public class LineItem : EntityCommon
-    {
-       
-        public double UnitAmount { get; set; }
-        public double Quantity { get; set; }
-        public double TaxableAmount { get; set; }
+    {       
+        public float UnitAmount { get; set; }
+        public float Quantity { get; set; }
+        public float TaxableAmount { get; set; }
         public string TaxType { get; set; } = "VAT";
-        public double TaxAmount { get; set; }   
+        public float TaxAmount { get; set; }   
         public string ElementCode { get; set; }
         public string VoucherCode { get; set; }
     }
