+++ Domain/Entities/Voucher.cs
@@ -8,8 +8,8 @@ namespace StockManagment.Domain.Entities
     {
         public string Type { get; set; }
         public DateTime TimeStamp { get; set; }
-        public double SubTotal { get; set; }
-        public double GrandTotal { get; set; }
+        public float SubTotal { get; set; }
+        public float GrandTotal { get; set; }
         public bool Void { get; set; }
         public string ConsigneeCode { get; set; }
     }
